albumFragment 분리
