+++ app/src/main/java/umc/study/umc_7th/Navigation.kt
@@ -1,9 +1,12 @@
 package umc.study.umc_7th
 
+import android.content.Context
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
 import java.lang.reflect.Modifier
 
 @Composable
@@ -15,3 +18,10 @@ public fun NavHost(
     builder: NavGraphBuilder.() -> Unit
 ): Unit{}
 
+//
+//class FakeNavController : PreviewParameterProvider<NavController>{
+//    override val values: Sequence<NavController>
+//        get(){
+//
+//        }
+//}
